event_simulator/requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
kafka-python==2.0.2
requests==2.25.0
urllib3==1.26.2